+++ evalml/model_understanding/prediction_explanations/explainers.py
@@ -6,6 +6,7 @@ import numpy as np
 import pandas as pd
 
 from evalml.exceptions import PipelineScoreError
+from evalml.model_family import ModelFamily
 from evalml.model_understanding.prediction_explanations._report_creator_factory import (
     _report_creator_factory
 )
@@ -22,7 +23,7 @@ def explain_predictions(pipeline, input_features, y, indices_to_explain, top_k_f
                         output_format="text"):
     """Creates a report summarizing the top contributing features for each data point in the input features.
 
-    XGBoost models and CatBoost multiclass classifiers are not currently supported.
+    XGBoost and Stacked Ensemble models, as well as CatBoost multiclass classifiers, are not currently supported.
 
     Arguments:
         pipeline (PipelineBase): Fitted pipeline whose predictions we want to explain with SHAP.
@@ -46,6 +47,8 @@ def explain_predictions(pipeline, input_features, y, indices_to_explain, top_k_f
     input_features = infer_feature_types(input_features)
     input_features = _convert_woodwork_types_wrapper(input_features.to_dataframe())
 
+    if pipeline.model_family == ModelFamily.ENSEMBLE:
+        raise ValueError(f"Pipeline must not be a stacked ensemble pipeline")
     if input_features.empty:
         raise ValueError("Parameter input_features must be a non-empty dataframe.")
     if output_format not in {"text", "dict", "dataframe"}:
@@ -68,7 +71,7 @@ def explain_predictions_best_worst(pipeline, input_features, y_true, num_to_expl
                                    include_shap_values=False, metric=None, output_format="text"):
     """Creates a report summarizing the top contributing features for the best and worst points in the dataset as measured by error to true labels.
 
-    XGBoost models and CatBoost multiclass classifiers are not currently supported.
+    XGBoost and Stacked Ensemble models, as well as CatBoost multiclass classifiers, are not currently supported.
 
     Arguments:
         pipeline (PipelineBase): Fitted pipeline whose predictions we want to explain with SHAP.
@@ -99,6 +102,8 @@ def explain_predictions_best_worst(pipeline, input_features, y_true, num_to_expl
     y_true = infer_feature_types(y_true)
     y_true = _convert_woodwork_types_wrapper(y_true.to_series())
 
+    if pipeline.model_family == ModelFamily.ENSEMBLE:
+        raise ValueError(f"Pipeline must not be a stacked ensemble pipeline")
     if not (input_features.shape[0] >= num_to_explain * 2):
         raise ValueError(f"Input features must be a dataframe with more than {num_to_explain * 2} rows! "
                          "Convert to a dataframe and select a smaller value for num_to_explain if you do not have "